834 - split tests for native private kernel into 3 different files
